pyproject.toml
[project]
name = "molgen-bench"
version = "0.1.0"
description = "Benchmark for 3D molecular generative models"
requires-python = ">=3.9"
dependencies = [
    # Core deep learning
    "torch>=2.0.0",

    # Scientific computing
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "pandas>=2.0.0",

    # Molecular chemistry
    "rdkit>=2023.3.1",
    "networkx>=3.1",

    # Configuration and utilities
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    "h5py>=3.8.0",

    # Visualization
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",

    # Machine learning utilities
    "scikit-learn>=1.3.0",

    # Optional: Uncomment for geometric features
    # "torch-geometric>=2.3.0",
    # "e3nn>=0.5.0",
]

[dependency-groups]
dev = [
    "pytest>=7.3.0",
    "pytest-cov>=4.1.0",
    "black>=23.3.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.setuptools.package-dir]
"" = "."
